Added inline styling to EVisa and renamed it ElectronicVisa
